summarizer-agent/requirements.txt
python-dotenv
langchain
google-search-results
openai
